move config to features/config
