+++ src/features/config/index.ts
@@ -1,0 +1,5 @@
+import { CONFIG_FALLBACK } from "./configFallback";
+import { loadConfig } from "./loadConfig";
+import { saveConfig } from "./saveConfig";
+
+export default {configFallback: CONFIG_FALLBACK, loadConfig, saveConfig}
